+++ splusdata/__init__.py
@@ -1,3 +1,3 @@
 from splusdata.connect import connect
 
-from splusdata.features import g_star, filterbw
+from splusdata.features import classification, filterbw
